do not build prompt for windows
