+++ commands/utils_windows.go
@@ -4,6 +4,7 @@
 package commands
 
 import (
+	"errors"
 	"fmt"
 	"os"
 )
@@ -18,6 +19,30 @@ func checkValidSocket(socketPath string) error {
 	}
 	if fi.Mode() != expectedSocketMode {
 		return fmt.Errorf("invalid file mode for file %q, it must be a socket with 0600 permissions", socketPath)
+	}
+
+	return nil
+}
+
+func getConfirmation(question string, dbConfirmation bool) error {
+	if err := checkInteractiveTerminal(); err != nil {
+		return fmt.Errorf("could not proceed, either enable --confirm flag or use an interactive shell to complete operation: %w", err)
+	}
+
+	var confirm string
+	if dbConfirmation {
+		fmt.Println("Have you performed a database backup? (YES/NO): ")
+		fmt.Scanln(&confirm)
+
+		if confirm != "YES" {
+			return errors.New("aborted: You did not answer YES exactly, in all capitals")
+		}
+	}
+
+	fmt.Println(question + " (YES/NO): ")
+	fmt.Scanln(&confirm)
+	if confirm != "YES" {
+		return errors.New("aborted: You did not answer YES exactly, in all capitals")
 	}
 
 	return nil
